scratch_gui-40: Changed the number of seconds from integer to double.

+++ src/main/java/artie/pedagogicalintervention/webservice/model/PedagogicalSoftwareData.java
@@ -26,7 +26,7 @@ public class PedagogicalSoftwareData {
 	private PedagogicalSoftwareDistance solutionDistance = new PedagogicalSoftwareDistance();
 	private LocalDateTime dateTime;
 	private boolean requestHelp;
-	private int secondsHelpOpen;
+	private double secondsHelpOpen;
 	private boolean finishedExercise;
 	private int validSolution;
 	private double grade;
@@ -90,8 +90,8 @@ public class PedagogicalSoftwareData {
 		this.requestHelp = requestHelp;
 	}
 
-	public int getSecondsHelpOpen(){return this.secondsHelpOpen;}
-	public void setSecondsHelpOpen(int secondsHelpOpen){this.secondsHelpOpen = secondsHelpOpen;}
+	public double getSecondsHelpOpen(){return this.secondsHelpOpen;}
+	public void setSecondsHelpOpen(double secondsHelpOpen){this.secondsHelpOpen = secondsHelpOpen;}
 
 	public boolean getFinishedExercise(){return this.finishedExercise;}
 	public void setFinishedExercise(boolean finishedExercise){this.finishedExercise = finishedExercise;}
@@ -151,7 +151,7 @@ public class PedagogicalSoftwareData {
 	 * @param screenShot
 	 */
 	public PedagogicalSoftwareData(StudentDTO student, Exercise exercise, PedagogicalSoftwareDistance solutionDistance, List<PedagogicalSoftwareElement> elements,
-								   boolean requestHelp, int secondsHelpOpen, boolean finishedExercise, int validSolution, double grade, Date lastLogin, String screenShot) {
+								   boolean requestHelp, double secondsHelpOpen, boolean finishedExercise, int validSolution, double grade, Date lastLogin, String screenShot) {
 		this.student = student;
 		this.exercise = exercise;
 		this.solutionDistance = solutionDistance;
